[refactor] post 패키지 구조 변경 (#223)
