+++ backend/src/main/java/dev/tripdraw/post/application/PostService.java
@@ -1,4 +1,4 @@
-package dev.tripdraw.application;
+package dev.tripdraw.post.application;
 
 import static dev.tripdraw.domain.file.FileType.POST_IMAGE;
 
@@ -6,19 +6,19 @@ import dev.tripdraw.application.file.FileUploader;
 import dev.tripdraw.domain.file.FileType;
 import dev.tripdraw.member.domain.Member;
 import dev.tripdraw.member.domain.MemberRepository;
-import dev.tripdraw.domain.post.Post;
-import dev.tripdraw.domain.post.PostCreateEvent;
-import dev.tripdraw.domain.post.PostRepository;
+import dev.tripdraw.post.domain.Post;
+import dev.tripdraw.post.domain.PostCreateEvent;
+import dev.tripdraw.post.domain.PostRepository;
 import dev.tripdraw.trip.domain.Point;
 import dev.tripdraw.trip.domain.Trip;
 import dev.tripdraw.trip.domain.TripRepository;
 import dev.tripdraw.dto.auth.LoginUser;
-import dev.tripdraw.dto.post.PostAndPointCreateRequest;
-import dev.tripdraw.dto.post.PostCreateResponse;
-import dev.tripdraw.dto.post.PostRequest;
-import dev.tripdraw.dto.post.PostResponse;
-import dev.tripdraw.dto.post.PostUpdateRequest;
-import dev.tripdraw.dto.post.PostsResponse;
+import dev.tripdraw.post.dto.PostAndPointCreateRequest;
+import dev.tripdraw.post.dto.PostCreateResponse;
+import dev.tripdraw.post.dto.PostRequest;
+import dev.tripdraw.post.dto.PostResponse;
+import dev.tripdraw.post.dto.PostUpdateRequest;
+import dev.tripdraw.post.dto.PostsResponse;
 import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
