+++ backend/src/main/java/dev/tripdraw/post/dto/PostRequest.java
@@ -1,7 +1,7 @@
-package dev.tripdraw.dto.post;
+package dev.tripdraw.post.dto;
 
 import dev.tripdraw.member.domain.Member;
-import dev.tripdraw.domain.post.Post;
+import dev.tripdraw.post.domain.Post;
 import dev.tripdraw.trip.domain.Point;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
