+++ backend/src/main/java/dev/tripdraw/post/dto/PostResponse.java
@@ -1,6 +1,6 @@
-package dev.tripdraw.dto.post;
+package dev.tripdraw.post.dto;
 
-import dev.tripdraw.domain.post.Post;
+import dev.tripdraw.post.domain.Post;
 import dev.tripdraw.trip.dto.PointResponse;
 import io.swagger.v3.oas.annotations.media.Schema;
