+++ backend/src/test/java/dev/tripdraw/member/presentation/MemberControllerTest.java
@@ -11,8 +11,8 @@ import static org.springframework.http.HttpStatus.OK;
 import dev.tripdraw.application.oauth.AuthTokenManager;
 import dev.tripdraw.member.domain.Member;
 import dev.tripdraw.member.domain.MemberRepository;
-import dev.tripdraw.domain.post.Post;
-import dev.tripdraw.domain.post.PostRepository;
+import dev.tripdraw.post.domain.Post;
+import dev.tripdraw.post.domain.PostRepository;
 import dev.tripdraw.member.dto.MemberSearchResponse;
 import dev.tripdraw.test.ControllerTest;
 import dev.tripdraw.trip.domain.Point;
