+++ backend/src/test/java/dev/tripdraw/post/application/PostServiceTest.java
@@ -1,9 +1,9 @@
-package dev.tripdraw.application;
+package dev.tripdraw.post.application;
 
 import static dev.tripdraw.domain.oauth.OauthType.KAKAO;
 import static dev.tripdraw.member.exception.MemberExceptionType.MEMBER_NOT_FOUND;
-import static dev.tripdraw.exception.post.PostExceptionType.NOT_AUTHORIZED_TO_POST;
-import static dev.tripdraw.exception.post.PostExceptionType.POST_NOT_FOUND;
+import static dev.tripdraw.post.exception.PostExceptionType.NOT_AUTHORIZED_TO_POST;
+import static dev.tripdraw.post.exception.PostExceptionType.POST_NOT_FOUND;
 import static dev.tripdraw.trip.exception.TripExceptionType.NOT_AUTHORIZED_TO_TRIP;
 import static dev.tripdraw.trip.exception.TripExceptionType.POINT_NOT_FOUND;
 import static dev.tripdraw.trip.exception.TripExceptionType.TRIP_NOT_FOUND;
@@ -14,21 +14,23 @@ import static org.assertj.core.api.SoftAssertions.assertSoftly;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.mockito.ArgumentMatchers.any;
 
+import dev.tripdraw.application.ServiceTest;
 import dev.tripdraw.draw.application.RouteImageGenerator;
 import dev.tripdraw.member.domain.Member;
 import dev.tripdraw.member.domain.MemberRepository;
+import dev.tripdraw.post.application.PostService;
 import dev.tripdraw.trip.domain.Point;
 import dev.tripdraw.trip.domain.Trip;
 import dev.tripdraw.trip.domain.TripRepository;
 import dev.tripdraw.dto.auth.LoginUser;
-import dev.tripdraw.dto.post.PostAndPointCreateRequest;
-import dev.tripdraw.dto.post.PostCreateResponse;
-import dev.tripdraw.dto.post.PostRequest;
-import dev.tripdraw.dto.post.PostResponse;
-import dev.tripdraw.dto.post.PostUpdateRequest;
-import dev.tripdraw.dto.post.PostsResponse;
+import dev.tripdraw.post.dto.PostAndPointCreateRequest;
+import dev.tripdraw.post.dto.PostCreateResponse;
+import dev.tripdraw.post.dto.PostRequest;
+import dev.tripdraw.post.dto.PostResponse;
+import dev.tripdraw.post.dto.PostUpdateRequest;
+import dev.tripdraw.post.dto.PostsResponse;
 import dev.tripdraw.member.exception.MemberException;
-import dev.tripdraw.exception.post.PostException;
+import dev.tripdraw.post.exception.PostException;
 import dev.tripdraw.trip.exception.TripException;
 import java.time.LocalDateTime;
 import java.util.List;
