+++ backend/src/test/java/dev/tripdraw/post/domain/PostTest.java
@@ -1,15 +1,16 @@
-package dev.tripdraw.domain.post;
+package dev.tripdraw.post.domain;
 
 import static dev.tripdraw.domain.oauth.OauthType.KAKAO;
-import static dev.tripdraw.exception.post.PostExceptionType.NOT_AUTHORIZED_TO_POST;
+import static dev.tripdraw.post.exception.PostExceptionType.NOT_AUTHORIZED_TO_POST;
 import static dev.tripdraw.trip.exception.TripExceptionType.POINT_ALREADY_HAS_POST;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import dev.tripdraw.member.domain.Member;
+import dev.tripdraw.post.domain.Post;
 import dev.tripdraw.trip.domain.Point;
-import dev.tripdraw.exception.post.PostException;
+import dev.tripdraw.post.exception.PostException;
 import dev.tripdraw.trip.exception.TripException;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayNameGeneration;
